Split plugin to use a hook and configure
